# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dcba54af1ef4518a5382e85
  title: Wishbox
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /category/:
    post:
      x-skaffolder-id: 5dcbc232f1ef4518a5382f41
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcba8c0f1ef4518a5382f09
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Category
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                desc:
                  type: string
                isPublic:
                  type: boolean
                name:
                  type: string
    get:
      x-skaffolder-id: 5dcbc232f1ef4518a5382f48
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcba8c0f1ef4518a5382f09
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Category
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /category/{id}:
    delete:
      x-skaffolder-id: 5dcbc232f1ef4518a5382f4a
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcba8c0f1ef4518a5382f09
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcbc232f1ef4518a5382f42
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcba8c0f1ef4518a5382f09
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    post:
      x-skaffolder-id: 5dcbc232f1ef4518a5382f43
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcba8c0f1ef4518a5382f09
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Category
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
  /category/findBycreateBy/{key}:
    get:
      x-skaffolder-id: 5dcbc22ef1ef4518a5382f39
      x-skaffolder-name: findBycreateBy
      x-skaffolder-id-resource: 5dcba8c0f1ef4518a5382f09
      x-skaffolder-resource: Category
      x-skaffolder-crudAction: findBycreateBy
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBycreateBy
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBycreateBy/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBycreateBy
      tags:
        - Category
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa createBy da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    
  #-----  Item  -----
  /item/:
    post:
      x-skaffolder-id: 5dcbc238f1ef4518a5382f50
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcba9bef1ef4518a5382f14
      x-skaffolder-resource: Item
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Item
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alias:
                  type: string
                createdAt:
                  type: integer
                desc:
                  type: string
                isPublic:
                  type: boolean
                name:
                  type: string
                numbers:
                  type: integer
                updatedAt:
                  type: integer
    get:
      x-skaffolder-id: 5dcbc238f1ef4518a5382f54
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcba9bef1ef4518a5382f14
      x-skaffolder-resource: Item
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Item
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /item/{id}:
    delete:
      x-skaffolder-id: 5dcbc238f1ef4518a5382f53
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcba9bef1ef4518a5382f14
      x-skaffolder-resource: Item
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Item
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcbc238f1ef4518a5382f4f
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcba9bef1ef4518a5382f14
      x-skaffolder-resource: Item
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Item
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    post:
      x-skaffolder-id: 5dcbc238f1ef4518a5382f4d
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcba9bef1ef4518a5382f14
      x-skaffolder-resource: Item
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Item
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
  /item/findBycategory/{key}:
    get:
      x-skaffolder-id: 5dcbc232f1ef4518a5382f46
      x-skaffolder-name: findBycategory
      x-skaffolder-id-resource: 5dcba9bef1ef4518a5382f14
      x-skaffolder-resource: Item
      x-skaffolder-crudAction: findBycategory
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBycategory
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBycategory/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBycategory
      tags:
        - Item
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa category da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /item/findBycreatedBy/{key}:
    get:
      x-skaffolder-id: 5dcbc22ef1ef4518a5382f3b
      x-skaffolder-name: findBycreatedBy
      x-skaffolder-id-resource: 5dcba9bef1ef4518a5382f14
      x-skaffolder-resource: Item
      x-skaffolder-crudAction: findBycreatedBy
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBycreatedBy
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBycreatedBy/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBycreatedBy
      tags:
        - Item
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa createdBy da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /item/findBytags/{key}:
    get:
      x-skaffolder-id: 5dcbc6a3f1ef4518a5382f95
      x-skaffolder-name: findBytags
      x-skaffolder-id-resource: 5dcba9bef1ef4518a5382f14
      x-skaffolder-resource: Item
      x-skaffolder-crudAction: findBytags
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBytags
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBytags/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBytags
      tags:
        - Item
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa tags da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    
  #-----  Post  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
    Category:
      x-skaffolder-id: 5dcba8c0f1ef4518a5382f09
      x-skaffolder-id-db: 5dcba54af1ef4518a5382e86
      x-skaffolder-id-entity: 5dcba8c0f1ef4518a5382f05
      x-skaffolder-url: /category
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        desc:
          type: string
          x-skaffolder-id-attr: 5dcba8c0f1ef4518a5382f07
          x-skaffolder-type: String
        isPublic:
          type: boolean
          x-skaffolder-id-attr: 5dcbc312f1ef4518a5382f65
          x-skaffolder-type: Boolean
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5dcba8c0f1ef4518a5382f06
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
        createBy:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dcba8c0f1ef4518a5382f05
          x-skaffolder-ent2: 5dcba54af1ef4518a5382e87
        parent:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dcba8c0f1ef4518a5382f05
          x-skaffolder-ent2: 5dcba8c0f1ef4518a5382f05
      required:
        - _id
        - isPublic
        - name
    Item:
      x-skaffolder-id: 5dcba9bef1ef4518a5382f14
      x-skaffolder-id-db: 5dcba54af1ef4518a5382e86
      x-skaffolder-id-entity: 5dcba9bdf1ef4518a5382f0c
      x-skaffolder-url: /item
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        alias:
          type: string
          x-skaffolder-id-attr: 5dcba9bdf1ef4518a5382f0e
          x-skaffolder-type: String
        createdAt:
          type: integer
          x-skaffolder-id-attr: 5dcba9bdf1ef4518a5382f10
          x-skaffolder-type: Date
        desc:
          type: string
          x-skaffolder-id-attr: 5dcba9bdf1ef4518a5382f0f
          x-skaffolder-type: String
        isPublic:
          type: boolean
          x-skaffolder-id-attr: 5dcbc367f1ef4518a5382f66
          x-skaffolder-type: Boolean
        name:
          type: string
          x-skaffolder-id-attr: 5dcba9bdf1ef4518a5382f0d
          x-skaffolder-type: String
          x-skaffolder-required: true
        numbers:
          type: integer
          x-skaffolder-id-attr: 5dcba9bdf1ef4518a5382f12
          x-skaffolder-type: Integer
        updatedAt:
          type: integer
          x-skaffolder-id-attr: 5dcba9bdf1ef4518a5382f11
          x-skaffolder-type: Date
      x-skaffolder-relations: 
        category:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dcba9bdf1ef4518a5382f0c
          x-skaffolder-ent2: 5dcba8c0f1ef4518a5382f05
        createdBy:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dcba9bdf1ef4518a5382f0c
          x-skaffolder-ent2: 5dcba54af1ef4518a5382e87
        tags:
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5dcba9bdf1ef4518a5382f0c
          x-skaffolder-ent2: 5dcbc504f1ef4518a5382f77
      required:
        - _id
        - name
...